+++ spec/integration/resource/api600/synergy/server_profile/delete_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::Synergy::ServerProfile
+klass = OneviewSDK::API600::Synergy::ServerProfile
 RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
-  let(:current_client) { $client_500_synergy }
-  include_examples 'ServerProfileDeleteExample', 'integration api500 context'
+  let(:current_client) { $client_600_synergy }
+  include_examples 'ServerProfileDeleteExample', 'integration api600 context'
 end