+++ CHANGELOG.md
@@ -25,6 +25,7 @@ This release adds support to OneView Rest API version 600 for the hardware varia
    - SAS Logical Interconnect Group
    - Server Hardware
    - Server Hardware Types
+   - Server Profile
    - Server Profile Template
    - Storage Pool
    - Storage System